+++ relstorage/tests/test_cache.py
@@ -403,45 +403,64 @@ class LocalClientBucketTests(unittest.TestCase):
         self.assertEqual(b.get('abc'), None)
         self.assertEqual(b.get("abcd"), 'xyz')
 
-    def test_load_and_store(self):
+    def _load(self, bio, bucket, options):
+        from relstorage.cache import _Loader
+        bio.seek(0)
+        reader = _Loader._gzip_file(options, None, bio, mode='rb')
+        return bucket.load_from_file(reader)
+
+    def _save(self, bio, bucket, options):
+        from relstorage.cache import _Loader
+        bio.seek(0)
+        if options.cache_local_dir_compress:
+            self.assertEqual(".rscache.gz", _Loader._gzip_ext(options))
+        writer = _Loader._gzip_file(options, None, bio, mode='wb')
+        bucket.write_to_file(writer)
+        writer.flush()
+        if writer is not bio:
+            writer.close()
+        bio.seek(0)
+        return bio
+
+    def test_load_and_store(self, options=None):
         from io import BytesIO
+        if options is None:
+            options = MockOptions()
         client1 = self.getClass()(100)
         client1['abc'] = b'xyz'
 
         bio = BytesIO()
-        client1.write_to_file(bio)
-        bio.seek(0)
+
+        self._save(bio, client1, options)
 
         client2 = self.getClass()(100)
-        count, stored = client2.load_from_file(bio)
+        count, stored = self._load(bio, client2, options)
         self.assertEqual(count, stored)
         self.assertEqual(count, 1)
         self.assertEqual(client1['abc'], client2['abc'])
         self.assertEqual(1, len(client2))
         self.assertEqual(client1.size, client2.size)
 
-        bio.seek(0)
         client1.reset_stats()
         client1['def'] = b'123'
+        self.assertEqual(2, len(client1))
         client1_max_size = client1.size
-        client1.write_to_file(bio)
-        bio.seek(0)
+        self._save(bio, client1, options)
 
         # This time there's too much data, so an arbitrary
         # entry gets dropped
         client2 = self.getClass()(3)
-        count, stored = client2.load_from_file(bio)
-        self.assertEqual(count, stored)
-        self.assertEqual(count, 2)
+        count, stored = self._load(bio, client2, options)
         self.assertEqual(1, len(client2))
+        self.assertEqual(count, 2)
+        self.assertEqual(stored, 1)
 
 
-        bio.seek(0)
         # Duplicate keys ignored.
         # Note that we do this in client1, because if we do it in client2,
         # the first key (abc) will push out the existing 'def' and get
         # inserted, and then 'def' will push out 'abc'
-        count, stored = client1.load_from_file(bio)
+        count, stored = self._load(bio, client1, options)
         self.assertEqual(count, 2)
         self.assertEqual(stored, 0)
         self.assertEqual(2, len(client1))
@@ -449,14 +468,43 @@ class LocalClientBucketTests(unittest.TestCase):
         # Half duplicate keys
         del client1['abc']
         self.assertEqual(1, len(client1))
-        bio.seek(0)
-        count, stored = client1.load_from_file(bio)
+
+        count, stored = self._load(bio, client1, options)
         self.assertEqual(client1['def'], b'123')
         self.assertEqual(client1['abc'], b'xyz')
         self.assertEqual(count, 2)
         self.assertEqual(stored, 1)
         self.assertEqual(client1.size, client1_max_size)
 
+    def test_load_and_store_to_gzip(self):
+        options = MockOptions()
+        options.cache_local_dir_compress = True
+        self.test_load_and_store(options)
+
+    def test_load_from_multiple_files_hit_limit(self):
+        from relstorage.cache import _Loader
+        import tempfile
+        client = self.getClass()(100)
+        options = MockOptions()
+        options.cache_local_dir_count = 5
+        options.cache_local_dir_read_count = 2
+        options.cache_local_dir = tempfile.mkdtemp()
+
+        for i in range(5):
+            # They all have to have unique keys so something gets loaded
+            # from each one
+            if i > 0:
+                del client[str(i - 1)]
+            client[str(i)] = b'abc'
+            _Loader.save_local_cache(options, 'test', client, _pid=i)
+            self.assertEqual(_Loader.count_cache_files(options, 'test'),
+                             i + 1)
+
+        files_loaded = _Loader.load_local_cache(options, 'test', client)
+        self.assertEqual(files_loaded, 2)
+
+        import shutil
+        shutil.rmtree(options.cache_local_dir)
 
 class LocalClientTests(unittest.TestCase):
 
@@ -660,8 +708,9 @@ class LocalClientTests(unittest.TestCase):
          # automatically create directories as needed
          self.test_load_and_save(False)
 
+from relstorage.options import Options
 
-class MockOptions(object):
+class MockOptions(Options):
     cache_module_name = ''
     cache_servers = ''
     cache_local_mb = 1
@@ -669,6 +718,7 @@ class MockOptions(object):
     cache_local_compression = 'zlib'
     cache_delta_size_limit = 10000
     cache_local_dir = None
+    cache_local_dir_compress = False
     cache_local_dir_count = 1
 
 class MockOptionsWithFakeCache(MockOptions):
@@ -694,111 +744,197 @@ class MockPoller(object):
                 if tid > after_tid and tid <= last_tid)
 
 def local_benchmark():
+    from relstorage.cache import LocalClient, LocalClientBucket
     options = MockOptions()
-    options.cache_local_mb = 100
+    options.cache_local_mb = 500
     #options.cache_local_compression = 'none'
 
-    from relstorage.cache import LocalClient
-    import time
-    client = LocalClient(options)
+    REPEAT_COUNT = 4
+
+    KEY_GROUP_SIZE = 400
+    DATA_SIZE = 1024
+
+    # With 1000 in a key group, and 1024 bytes of data, we produce
+    # 909100 keys, and 930918400 = 887MB of data, which will overflow
+    # a cache of 500 MB.
+
+    # A group size of 100 produces 9100 keys with 9318400 = 8.8MB of data.
+    # Likewise, group of 200 produces 36380 keys with 35.5MB of data.
+
+    # Most of our time is spent in compression, it seems.
+    # In the 8.8mb case, populating all the data with default compression
+    # takes about 2.5-2.8s. Using no compression, it takes 0.38 to 0.42s.
+    # Reading is the same at about 0.2s.
+
     with open('/dev/urandom', 'rb') as f:
-        random_data = f.read(1024)
+        random_data = f.read(DATA_SIZE)
 
     key_groups = []
-    key_groups.append([str(i) for i in range(10)])
-    for i in range(1, 10): # 1 - 9
-        keys = [str(i) + str(j) for j in range(10)]
+    key_groups.append([int(str(i)) for i in range(KEY_GROUP_SIZE)])
+    for i in range(1, KEY_GROUP_SIZE):
+        keys = [int(str(i) + str(j)) for j in range(KEY_GROUP_SIZE)]
         key_groups.append(keys)
 
-    def populate():
-        data = {str(k): random_data for k in range(120)}
-        for k, v in data.items():
-            client.set(k, v)
+    ALL_DATA = {}
+    for group in key_groups:
+        for key in group:
+            ALL_DATA[key] = random_data
+    print(len(ALL_DATA), sum((len(v) for v in ALL_DATA.values()))/1024/1024)
 
-    def read():
-        for keys in key_groups:
-            client.get_multi(keys)
+    class DLocalBucket(LocalClientBucket):
+        CACHE_TYPE = dict
 
-    import timeit
-    import statistics
-    #import cProfile, pstats
-    number = 100
-    pop_timer = timeit.Timer(populate)
-    #pr = cProfile.Profile()
-    #pr.enable()
-    pop_times = pop_timer.repeat(number=number)
-    #pr.disable()
-    #ps = pstats.Stats(pr).sort_stats('cumulative')
-    #ps.print_stats()
+    class DLocalClient(LocalClient):
+        bucket_type = DLocalBucket
 
-    read_timer = timeit.Timer(read)
-    #pr = cProfile.Profile()
-    #pr.enable()
-    read_times = read_timer.repeat(number=number)
-    #pr.disable()
-    #ps = pstats.Stats(pr).sort_stats('cumulative')
-    #ps.print_stats()
+    from BTrees.OOBTree import OOBTree
+    from BTrees.LOBTree import LOBTree
+    class BLocalBucket(LocalClientBucket):
+        CACHE_TYPE = OOBTree
+
+    class BLocalClient(LocalClient):
+        bucket_type = BLocalBucket
+
+    def do_times(client_type):
+        client = client_type(options)
+        print("Testing", type(client._bucket0._dict))
+
+        def populate():
+            for k, v in ALL_DATA.items():
+                client.set(k, v)
+
+        def populate_empty():
+            c = LocalClient(options)
+            for k, v in ALL_DATA.items():
+                c.set(k, v)
+
+        def read():
+            for keys in key_groups:
+                res = client.get_multi(keys)
+                assert len(res) == len(keys)
+                assert res.popitem()[1] == random_data
 
 
-    print("pop average", statistics.mean(pop_times), "stddev", statistics.stdev(pop_times))
-    print("read average", statistics.mean(read_times), "stddev", statistics.stdev(read_times))
+
+        import timeit
+        import statistics
+        try:
+            import cProfile, pstats
+            raise ImportError
+        except ImportError:
+            class cProfile(object):
+                class Profile(object):
+                    def enable(self): pass
+                    def disable(self): pass
+            class pstats(object):
+                class Stats(object):
+                    def __init__(self, *args): pass
+                    def sort_stats(self, *args): return self
+                    def print_stats(self, *args): pass
+
+
+        number = REPEAT_COUNT
+        pop_timer = timeit.Timer(populate)
+        pr = cProfile.Profile()
+        pr.enable()
+        pop_times = pop_timer.repeat(number=number)
+        pr.disable()
+        ps = pstats.Stats(pr).sort_stats('cumulative')
+        ps.print_stats(.4)
+
+        read_timer = timeit.Timer(read)
+        pr = cProfile.Profile()
+        pr.enable()
+        read_times = read_timer.repeat(number=number)
+        pr.disable()
+        ps = pstats.Stats(pr).sort_stats('cumulative')
+        ps.print_stats(.4)
+
+        empty_pop = timeit.Timer(populate_empty)
+        epop_times = empty_pop.repeat(number=number)
+
+        print("pop  average", statistics.mean(pop_times), "stddev", statistics.stdev(pop_times))
+        print("epop average", statistics.mean(epop_times), "stddev", statistics.stdev(epop_times))
+        print("read average", statistics.mean(read_times), "stddev", statistics.stdev(read_times))
+
+    do_times(DLocalClient)
+    do_times(BLocalClient)
 
 def save_load_benchmark():
-    from relstorage.cache import LocalClientBucket
+    from relstorage.cache import LocalClientBucket, _Loader
     from io import BytesIO
+    import os
+    import itertools
 
-    bucket = LocalClientBucket(10*1024*1024)
+    import sys
+    sys.setrecursionlimit(500000)
+    bucket = LocalClientBucket(500*1024*1024)
+    print("Testing", type(bucket._dict))
 
-    i = 1
-    j = 0
-    while bucket.size < bucket.limit - i:
-        val = (str(j) * i).encode('ascii')
-        if len(val) > bucket.limit or bucket.size + len(val) > bucket.limit:
+
+    size_dists = [100] * 800 + [300] * 500 + [1024] * 300 + [2048] * 200 + [4096] * 150
+
+    with open('/dev/urandom', 'rb') as rnd:
+        data = [rnd.read(x) for x in size_dists]
+    data_iter = itertools.cycle(data)
+
+    for j, datum in enumerate(data_iter):
+        if len(datum) > bucket.limit or bucket.size + len(datum) > bucket.limit:
             break
-        bucket[str(i)] = val
-        if i < 1096:
-            i += 50
-            j += 1
-        else:
-            j += 1
-            i += 1
-        print("Len", len(bucket), "size",bucket.size, "i", i)
+        # To ensure the pickle memo cache doesn't just write out "use object X",
+        # but distinct copies of the strings, we need to copy them
+        bucket[str(j)] = datum[:-1] + b'x'
+        assert bucket[str(j)] is not datum
+        #print("Len", len(bucket), "size", bucket.size, "dlen", len(datum))
 
     print("Len", len(bucket), "size", bucket.size)
-    number = 50
+    number = 1
     import timeit
     import statistics
     import cProfile
     import pstats
 
-    def write():
-        io = BytesIO()
-        bucket.write_to_file(io)
+    cache_pfx = "pfx"
+    cache_options = MockOptions()
+    cache_options.cache_local_dir = '/tmp'
+    cache_options.cache_local_dir_compress = False
 
-    bio = BytesIO()
-    bucket.write_to_file(bio)
+    def write():
+        fname = _Loader.save_local_cache(cache_options, cache_pfx, bucket)
+        os.remove(fname)
+
 
     def load():
-        bio.seek(0)
         b2 = LocalClientBucket(bucket.limit)
-        b2.load_from_file(bio)
+        _Loader.load_local_cache(cache_options, cache_pfx, b2)
 
-    write_timer = timeit.Timer(write)
-    write_times = write_timer.repeat(number=number)
+    #write_timer = timeit.Timer(write)
+    #write_times = write_timer.repeat(number=number)
+    #print("write average", statistics.mean(write_times), "stddev", statistics.stdev(write_times))
 
-    print("write average", statistics.mean(write_times), "stddev", statistics.stdev(write_times))
+    #read_timer = timeit.Timer(load)
+    #read_times = read_timer.repeat(number=number)
+    #print("read average", statistics.mean(read_times), "stddev", statistics.stdev(read_times))
 
     #pr = cProfile.Profile()
     #pr.enable()
 
-    read_timer = timeit.Timer(load)
-    read_times = read_timer.repeat(number=number)
+    fname = _Loader.save_local_cache(cache_options, cache_pfx, bucket)
+    print("Saved to", fname)
     #pr.disable()
     #ps = pstats.Stats(pr).sort_stats('cumulative')
     #ps.print_stats()
+    #return
 
-    print("read average", statistics.mean(read_times), "stddev", statistics.stdev(read_times))
+    pr = cProfile.Profile()
+    pr.enable()
+    _Loader.load_local_cache(cache_options, cache_pfx, LocalClientBucket(bucket.limit))
+    pr.disable()
+    ps = pstats.Stats(pr).sort_stats('cumulative')
+    ps.print_stats(.4)
 
+
+    #os.remove(fname)
 
 def test_suite():
     suite = unittest.TestSuite()
@@ -812,6 +948,8 @@ if __name__ == '__main__':
     if '--localbench' in sys.argv:
         local_benchmark()
     elif '--iobench' in sys.argv:
+        import logging
+        logging.basicConfig(level=logging.DEBUG)
         save_load_benchmark()
     else:
         unittest.main(defaultTest='test_suite')